Robust int type
WebssInt is now exactly 64 bits long (it was too before, but now its type
uses a standard constant).

There was an error in the deserializer where the assert for WebssInt
checked the min of "signed int" rather than that of WebssInt.

Another point has to be made. For Webss constructors, the validity of ints
and doubles should not be checked. That is because binary values use the
same types, and their values are allowed to be invalid. Only during
non-binary deserialization is their validity checked.

+++ WebssonStructures/types.h
@@ -3,9 +3,10 @@
 #pragma once
 
 #include <string>
+#include <cstdint>
 
 namespace webss
 {
-	using WebssInt = long long int;
+	using WebssInt = int64_t;
 	using WebssBinarySize = std::string::size_type;
 }
